refactor: reorganize file structure and update data_packet_t payload type

+++ transmitter/components/wcan/inc/wcan_communication.h
@@ -21,7 +21,7 @@ typedef struct {
     uint8_t attempt_number;
     uint8_t data_count;
     uint8_t *payload;
-    int payload_len;
+    uint8_t payload_len;
 } data_packet_t;
 
 typedef struct {
